OW fixes for ER
Lots of entrances needed to have their names slightly redone. Also needed to add some new entrances.

A few of these entrances are completely useless in glitch logic but still need to exist anyway since I don't want to need to change the names conditionally in the ER code because dear god that'd be dumb.