Glitched: adjust hovering in Colossus
Requires "All Uses Enabled" for hovering to the PoH and glitched damage
hovering from the arch to the hands without an item for contortion
hovering because of the length of these hovers.

I accounted for empty jumpslash instead of QPA for glitched damage,
which removes requirements for sticks and a cutscene item. Glitched
damage hovering now requires bombs. Glitched damage hovering to the
overhang requires both bombs and chus: bombs for hovering and a chu to
unlock the camera.

For brevity I removed hovers to Silver Gauntlets Hand that are redundant
with hovering to Mirror Shield Hand and then to Silvers.

Bombchu hovering without hoverboots is now expected when hoverboots
don't provide significant benefit.

I added a comment with how many explosives variations of hovering from
the arch to Mirror Shield Hand takes to document the exception to the
10 explosive limit.