Glitched: update logic for reaching Twinrova
Spirit Temple boss door and BK skip logic. The new logic considers
statue climb and head clip with no additional items and boss key skip by
Iron Boots megaflip or hoverboots.

Do not expect a damage ground clip BK skip when boss rooms are shuffled.
With damage multipliers it could leave insufficient health to defeat
Volvagia with a fairy, and there could be other issues if mixed pools
boss shuffle is added.

Also changed the exit for the elevator from the top to central chamber
to lead to adult side and removed skipping the elevator with hookshot.
It might've been expecting a weirdshot, but there's nothing in hookshot
range. The connection was redundant in the original logic but not
anymore with the hammerboosts in Colossus.

With these changes, the boss door access from the top is redundant, so
I removed the connection.

Added a description to the Spirit BK skip glitch toggle.