Glitched: revise Early Child Spirit
Split the part across the gaps into a new region. Rang was in logic as
an alternative to nuts for stunning the Armos. I removed that because
child has nuts available from the crate. Also, I think it was assumed
that adult would already have an item for crossing the gaps. But with
the TSC+crawlspace in logic, that's no longer the case; fixed. I added
restrictions for the nut crate as adult and duplicated it to Child
Climb with alternative restrictions so that it can be gotten by coming
from Central Chamber.

Add missing item and glitch toggle requirements to skip the crawlspaces.
I'm guessing the 2nd crawlspace with hoverboots is supposed to be EPG
and hookshot clip. I added the possibility of EPG there with explosives.
Also added EPG to the description of the actor glitch/EPG glitch toggle.