Glitched: add alternatives to damage in Spirit
After killing the child climb GS with a projectile, roll after jumping
to the token to avoid fall damage. Can also hover off of it or do an
iron boots dive (new glitch toggle).

For BK Chest, opening it with the flames up only loses a quarter heart
if you have roll invincibility for the first hit and put out the flames
by slashing. Added fairy or hookshot+nayrus to open the flaming chest,
slingshot to spawn the platform, and hovering to reach the switch
without the platform.

I considered removing the damage logic for BK Chest so that OHKO players
could be required to get a game over after getting the item, but I
figure that'd be a little tedious, so it's not worth doing.