Glitched: detrickify iron boots dive
Only used for 1 GS in Spirit with OHKO, and cutscene dives are generally
in base logic. When I added it, I thought it'd also be useful for LH GS
Lab Crate, but for that, you can also do the same ISG cutscene dive
that's in logic for Lab Dive, except with iron boots equipped.